refactor gcode and edge
